chore: move seeder out of prod code
