+++ lab-server/package.json
@@ -47,8 +47,7 @@
         "rimraf": "^3.0.2",
         "rxjs": "^7.2.0",
         "swagger-ui-express": "^4.2.0",
-        "typeorm": "^0.3.5",
-        "typeorm-seeding": "^1.6.1"
+        "typeorm": "^0.3.5"
     },
     "devDependencies": {
         "@commitlint/cli": "^17.3.0",
@@ -69,6 +68,7 @@
         "eslint-plugin-import": "^2.26.0",
         "eslint-plugin-nestjs": "^1.2.3",
         "eslint-plugin-prettier": "^4.0.0",
+        "fs": "^0.0.1-security",
         "jest": "^27.2.5",
         "keycloak-mock": "^1.0.10",
         "prettier": "^2.6.2",
