+++ lab-server/seeder/.gitignore
@@ -1,0 +1,1 @@
+users.json
